Performance improvements
slim-cover no longer uses Falafel.  Using Falafel was a great way to
bootstrap the project, but it also hindered it in a few ways.

* It imports the parser of its choosing.  All I could do was update to
the latest version via npm-shrinkwrap, which was not fun.  We're now
using Acorn, which is currently a bit faster.

* It walks the AST in an arbitrary order.  I had to sort the statements
by starting character index in order to keep the shell logic simple.  By
contrast, Traveler walks the AST in order.

* The AST walking code is a little slow.  It iterates over the keys of
each node object and expands all non-null elements belonging to child
objects and arrays.  By contrast, Traveler leverages the ESTree
specification to expand the child nodes relevant to each node type.

* The update() implementation keeps me awake at night.  I don't know if
it's slow.  All I know is there has to be a better way.  And that way is
Transformer.

This commit also fixes an instrumentation issue where For..Of loop
bodies are not wrapped in block statements if necessary.