Remove dependencies on supported test frameworks
This reduces the size of the module by assuming that the user has
already installed their test framework either locally or globally, which
is true in most if not all cases.

The real drawback is that slim-cover doesn't require a particular
supported version of the test framework, so if a change is made to one
of them that breaks the hook used by slim-cover, slim-cover would fail
to cover those tests.