Move coverage data out of global scope
slim-cover now sets no global variables.  Doing it this way didn't seem
to work before but works fine now.

slim-cover can also now cover itself, though this reminds me that it
only considers files that have been required by the tests, and not the
entirety of your source tree.  Because of this the project thinks it has
100% code coverage, when in reality it only has 100% coverage on
core/Instrumenter.js and 0% on everything else...